Cursor_First/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# 数据库相关
pymysql==1.1.0
redis==5.0.1
pymilvus==2.3.4

# AI/ML相关
torch==2.1.1
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# 向量数据库
faiss-cpu==1.7.4

# 工具库
requests==2.31.0
python-multipart==0.0.6
loguru==0.7.2
tiktoken==0.5.2

# 大模型接口
openai==1.3.7


